feat(app): move SearchView to features/noteSearch feat(noteSearch): create SearchSuggestion component feat(noteSearch): create SearchView component feat(noteSearch): implement search functionality feat(noteSearch): implement search suggestion feat(noteSearch): implement search result display feat(noteSearch): implement search result click feat(noteSearch): implement search clear button feat(noteSearch): implement search loading state feat(noteSearch): implement search empty state feat(noteSearch): implement search by tag feat(noteSearch): implement search by path feat(noteSearch): implement search by file refactor(main): remove SearchView import from ui
